+++ src/Http/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The change log describes what is "Added", "Removed", "Changed" or "Fixed" between each release.
 
-## 4.0.0
+## 4.0.0-beta2
+
+- Removed `AbstractHttpProvider::setMessageFactory`.
+- Removed `AbstractHttpProvider::getHttpClient`.
+- Make sure we have a `MessageFactory` in the constructor of `AbstractHttpProvider`. 
+
+## 4.0.0-beta1
 
 First release of this library.